chore: move compare data class to own file
