+++ src/app/compare/compare-data.ts
@@ -1,0 +1,9 @@
+import { NodeLinkStrategy } from '../shared/enums/compare-method';
+
+export class CompareData {
+  id!: string;
+  originalReport: any;
+  runResultReport: any;
+  viewName: string = '';
+  nodeLinkStrategy: NodeLinkStrategy = NodeLinkStrategy.NONE;
+}
