+++ src/app/text-compare/text-compare.component.html
@@ -1,2 +1,4 @@
-<ngx-monaco-diff-editor class="diff-editor" [options]="diffOptions" [originalModel]="originalModel"
-                        [modifiedModel]="modifiedModel"></ngx-monaco-diff-editor>
+<div *ngIf="compareData">
+  <ngx-monaco-diff-editor class="diff-editor" [options]="diffOptions" [originalModel]="originalModel"
+                          [modifiedModel]="modifiedModel"></ngx-monaco-diff-editor>
+</div>
