+++ src/app/text-compare/text-compare.component.ts
@@ -1,6 +1,12 @@
-import { Component, Input, OnInit, Output, ViewEncapsulation } from '@angular/core';
+import {
+  Component,
+  Input,
+  OnInit,
+  Output,
+  ViewEncapsulation,
+} from '@angular/core';
 import { DiffEditorModel } from 'ngx-monaco-editor-v2';
-import { CompareData } from '../compare/compare.component';
+import { CompareData } from '../compare/compare-data';
 
 @Component({
   selector: 'app-text-compare',
@@ -14,54 +20,6 @@ export class TextCompareComponent implements OnInit {
 
   @Output() selectedLang = 'plaintext';
   @Output() selectedTheme = 'vs';
-
-  @Input() languages: string[] = [
-    'bat',
-    'c',
-    'coffeescript',
-    'cpp',
-    'csharp',
-    'csp',
-    'css',
-    'dockerfile',
-    'fsharp',
-    'go',
-    'handlebars',
-    'html',
-    'ini',
-    'java',
-    'javascript',
-    'json',
-    'less',
-    'lua',
-    'markdown',
-    'msdax',
-    'mysql',
-    'objective-c',
-    'pgsql',
-    'php',
-    'plaintext',
-    'postiats',
-    'powershell',
-    'pug',
-    'python',
-    'r',
-    'razor',
-    'redis',
-    'redshift',
-    'ruby',
-    'rust',
-    'sb',
-    'scss',
-    'sol',
-    'sql',
-    'st',
-    'swift',
-    'typescript',
-    'vb',
-    'xml',
-    'yaml',
-  ];
 
   @Input() themes: { value: string; name: string }[] = [
     {
@@ -103,12 +61,14 @@ export class TextCompareComponent implements OnInit {
   }
 
   onCompare() {
-    this.originalModel = Object.assign({}, this.originalModel, {
-      code: this.compareData.originalReport.xml,
-    });
-    this.modifiedModel = Object.assign({}, this.originalModel, {
-      code: this.compareData.runResultReport.xml,
-    });
-    this.isCompared = true;
+    if (this.compareData) {
+      this.originalModel = Object.assign({}, this.originalModel, {
+        code: this.compareData.originalReport.xml,
+      });
+      this.modifiedModel = Object.assign({}, this.originalModel, {
+        code: this.compareData.runResultReport.xml,
+      });
+      this.isCompared = true;
+    }
   }
 }
